login register chpterid_Twila

+++ Server/E_learningAppBack/Models/Chapter.cs
@@ -13,7 +13,7 @@ namespace E_learningAppBack.Models
 
         public int ChapterId { get; set; }
         public string ChapterName { get; set; } = null!;
-        public int ScoreChapter { get; set; }
+        public double ScoreChapter { get; set; }
 
         public virtual ICollection<Course> Courses { get; set; }
         public virtual ICollection<Quiz> Quizzes { get; set; }
